refactor: text imports are now via mixins
